requirements.txt
deepface==0.0.68
h5py==3.1.0
matplotlib==3.4.3
numpy==1.19.5
opencv_python==4.5.4.58
pandas==1.3.3
Pillow==8.4.0
streamlit==1.1.0
torch==1.10.0
torchvision==0.11.1
